Rename tools folder to edit
