+++ visualizer/src/Project/service/edit/editCells/deleteMachine.js
