+++ visualizer/src/Project/service/edit/segment/thresholdMachine.js
