+++ visualizer/src/Project/service/edit/segmentMachine.js
